Add support for standard ingress relation
